added additional checkin test
