+++ __mocks__/mocks.js
@@ -7,39 +7,11 @@ imports
 
 import mockReactNativePermissions from 'react-native-permissions/mock'
 import mockAsyncStorage from '@react-native-community/async-storage/jest/async-storage-mock'
-import store from '../src/store.js';
-import MockAsyncStorage from 'mock-async-storage';
-import s from '../__utils__/mockStore'
-import { buildStore } from '../__utils__/mockStore'
-import _store from '../__utils__/mockStore'
-
-/***********************************************************************************************
-mock values
-***********************************************************************************************/
-
-// let mockState = {
-//     Login: {},
-//     CheckIn: {},
-//     About: {
-//         currentWebView: {}
-//     }
-// }
-
-mockState = buildStore().getState()
-
-
-/***********************************************************************************************
-preparation
-***********************************************************************************************/
-console.log(_store)
-// store.getState is going to be mocked
-store.getState = () => _store.getState()
+import MockAsyncStorage from 'mock-async-storage'
 
 // mocks the async storage
 const mockImpl = new MockAsyncStorage()
 
-
-jest.mock('../src/store.js')
 jest.mock('@react-native-community/async-storage', () => mockAsyncStorage)
 jest.mock('react-native-permissions', () => mockReactNativePermissions)
 jest.mock('@react-native-community/datetimepicker', () => jest.fn())
